+++ src/router/Router.tsx
@@ -19,8 +19,8 @@ import PlaceMake2 from '../pages/MakePlace/PlaceMake2';
 import PlaceMake3 from '../pages/MakePlace/PlaceMake3';
 import PlaceJoin1 from '../pages/JoinPlace/PlaceJoin1';
 import PlaceJoin2 from '../pages/JoinPlace/PlaceJoin2';
-import SettingManager from '../pages/I/SettingManager';
-import SettingMember from '../pages/I/SettingMember';
+import SettingManager from '../pages/Setting/SettingManager';
+import SettingMember from '../pages/Setting/SettingMember';
 import ManagementManager from '../pages/Dangbun/ManagementManager';
 import ManagementMember from '../pages/Dangbun/ManagementMember';
 import DutyMember from '../pages/Dangbun/DutyMember';
@@ -35,12 +35,12 @@ import UnDangbun from '../pages/Cleaning/UnDangbun';
 import CleanInfo from '../pages/Cleaning/CleanInfo';
 import CleanUpList_Member from '../pages/Cleaning/CleanUpList_Member';
 import UnDangbun_Member from '../pages/Cleaning/UnDangbun_Member';
-import ManagerInfo from '../pages/L/ManagerInfo';
-import MemberConfirm from '../pages/L/MemberConfirm';
-import MemberList from '../pages/L/MemberList';
-import EnterCode from '../pages/L/EnterCode';
-import MemberList_Member from '../pages/L/MemberList_Member';
-import ManagerInfo_Member from '../pages/L/ManagerInfo_Member';
+import ManagerInfo from '../pages/Member/ManagerInfo';
+import MemberConfirm from '../pages/Member/MemberConfirm';
+import MemberList from '../pages/Member/MemberList';
+import EnterCode from '../pages/Member/EnterCode';
+import MemberList_Member from '../pages/Member/MemberList_Member';
+import ManagerInfo_Member from '../pages/Member/ManagerInfo_Member';
 import PlaceDetailed from '../pages/Time/PlaceDetailed';
 import DangerZoneManager from '../pages/DangerZone/DangerZoneManager';
 import DangerZoneMember from '../pages/DangerZone/DangerZoneMember';
